+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Publicly access video or document, bypassing LTI check
 
+### Fixed
+
+- revert removal of mediaconvert presets configuration
+
 ## [3.13.1] - 2020-12-15
 
 ### Fixed